cdp_data_retriever/requirements.txt
xarray
cdsapi>=0.7.2
google-cloud-storage==2.18.2